More changes to Parser

+++ TabToMXL/src/main/java/Models/Bar.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 public class Bar {
     public final List<Note> notes;
-    public final int beatsPerBar;
+    public final long beatsPerBar;
 
-    public Bar(List<Note> notes, int beatsPerBar) {
+    public Bar(List<Note> notes, long beatsPerBar) {
         this.notes = notes;
         this.beatsPerBar = beatsPerBar;
     }
